fix:clean base template for creating credential
